Finished Drivetrain Stuff ( UNTESTED )
I finished the basic drivetrain controls and such. Will have to see if
it works. I moved around the OI initialization because it needs to be
the last thing out of the subsystems intialized. I also moved drivetrain
stuff into the teleop periodic instead of command based. Command based
implementation blocks the drivetrain if we hit anything else. This
should fix that.